Moved e2e related plugin/tasks to appropriate package
