+++ src/main/resources/META-INF/gradle-plugins/org.shipkit.e2e-test.properties
@@ -1,1 +1,1 @@
-implementation-class=org.shipkit.internal.gradle.E2ETestingPlugin
+implementation-class=org.shipkit.internal.gradle.e2e.E2ETestingPlugin
